+++ tests/model-client.js
@@ -13,6 +13,14 @@ describe('ClientModel', () => {
 
 	describe('Getters', () => {
 
+		it('Should return collection name when table getter is used', () => {
+			assert.equal(ClientModel.table, 'clients');
+		});
+
+		it('Should return indexes when indexes getter is used', () => {
+			assert(Array.isArray(ClientModel.indexes));
+		});
+
 		it('Should return the uniqueIndexes', async () => {
 			assert.deepStrictEqual(ClientModel.uniqueIndexes, ['code']);
 		});
@@ -24,5 +32,6 @@ describe('ClientModel', () => {
 		it('Should return false when shouldCreateLogs', () => {
 			assert.deepStrictEqual(ClientModel.shouldCreateLogs, false);
 		});
+
 	});
 });